# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales/goals/config/goal-csv-join-fields:
    put:
      operationId: updateGoalsGoalCsvJoinFields
      tags: [sales]
      summary: Replace the Goal CSV → hierarchy join fields
      description: Restricted to admin / SLM / exec and blocked for view-only sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: array
                  items:
                    $ref: "#/components/schemas/GoalCsvJoinField"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                required: [value]
                properties:
                  value:
                    type: array
                    items:
                      $ref: "#/components/schemas/GoalCsvJoinField"
        "400":
          $ref: "#/components/responses/GoalsValidationError"
...
  /sales/goals/config/goal-csv-inspect-columns:
    put:
      operationId: updateGoalsGoalCsvInspectColumns
      tags: [sales]
      summary: Replace the Goal CSV inspect column selection
      description: Restricted to admin / SLM / exec and blocked for view-only sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                required: [value]
                properties:
                  value:
                    type: array
                    items:
                      type: string
        "400":
          $ref: "#/components/responses/GoalsValidationError"
...
  /sales/goals/config/finance-pps-join-fields:
    put:
      operationId: updateGoalsFinancePpsJoinFields
      tags: [sales]
      summary: Replace the finance.pps → hierarchy join fields
      description: Restricted to admin / SLM / exec and blocked for view-only sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: array
                  items:
                    $ref: "#/components/schemas/FinancePpsJoinField"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                required: [value]
                properties:
                  value:
                    type: array
                    items:
                      $ref: "#/components/schemas/FinancePpsJoinField"
        "400":
          $ref: "#/components/responses/GoalsValidationError"
...
  /sales/goals/config/finance-pps-inspect-columns:
    put:
      operationId: updateGoalsFinancePpsInspectColumns
      tags: [sales]
      summary: Replace the finance.pps inspect column selection
      description: Restricted to admin / SLM / exec and blocked for view-only sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                required: [value]
                properties:
                  value:
                    type: array
                    items:
                      type: string
        "400":
          $ref: "#/components/responses/GoalsValidationError"
  /sales/goals/config/software-gnr-rules:
    put:
      operationId: updateGoalsSoftwareGnrRules
      tags: [sales]
      summary: Replace the Software % GNR rules config
      description: Restricted to admin / SLM / exec and blocked for view-only sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  $ref: "#/components/schemas/SoftwarePctRules"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                required: [value]
                properties:
                  value:
                    $ref: "#/components/schemas/SoftwarePctRules"
        "400":
          $ref: "#/components/responses/GoalsValidationError"
  /sales/goals/config/software-acq-rules:
    put:
      operationId: updateGoalsSoftwareAcqRules
      tags: [sales]
      summary: Replace the Software % ACQ rules config
      description: Restricted to admin / SLM / exec and blocked for view-only sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  $ref: "#/components/schemas/SoftwarePctRules"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                required: [value]
                properties:
                  value:
                    $ref: "#/components/schemas/SoftwarePctRules"
        "400":
          $ref: "#/components/responses/GoalsValidationError"
...
  /sales/goals/goal-csv/upload:
    post:
      operationId: uploadGoalsGoalCsv
      tags: [sales]
      summary: Replace all Goal CSV rows from an uploaded CSV
      description: |
        Parses the CSV (every column is preserved verbatim; no required headers)
        then replaces all stored rows. Restricted to admin / SLM / exec and
        blocked for view-only sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [csv]
              properties:
                csv:
                  type: string
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                type: object
                required: [inserted]
                properties:
                  inserted:
                    type: integer
        "400":
          $ref: "#/components/responses/GoalsValidationError"
  /sales/goals/resolve:
    get:
      operationId: resolveGoals
      tags: [sales]
      summary: Resolve goals for one source for a month
      description: |
        Resolves a single goal source into a per-rep, per-product goal set for
        the given month (`YYYY-MM`, defaults to current). Products/metrics a
        source does not populate are absent (zero downstream). Restricted to
        admin / SLM / exec.
      parameters:
        - name: source
          in: query
          required: true
          schema:
            type: string
            enum: [financePps, goalCsv, softwareGnr, softwareAcq]
        - name: month
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        "200":
          description: Resolved goals
          content:
            application/json:
              schema:
                type: object
                required: [source, month, goals]
                properties:
                  source:
                    type: string
                    enum: [financePps, goalCsv, softwareGnr, softwareAcq]
                  month:
                    type: string
                  goals:
                    $ref: "#/components/schemas/ResolvedGoals"
        "400":
          $ref: "#/components/responses/GoalsValidationError"
  /sales/goals/table:
    get:
      operationId: getGoalTable
      tags: [sales]
      summary: Get the main Goals table for the dashboard filter set
      description: |
        Returns one row per non-SLM/FLM rep × product across the five canonical
        products for `month` (`YYYY-MM`, defaults to current), narrowed by the
        optional dashboard filters. Each row carries its hierarchy attributes,
        its Source (default finance.pps) and the base goals that source resolves
        to, the per-row overrides (manual multipliers, LOA status, eRep
        multiplier), and the four computed Final goal columns. Restricted to
        admin / SLM / exec.
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
        - name: slm
          in: query
          required: false
          description: Filter to a single SLM by name.
          schema:
            type: string
        - name: flm
          in: query
          required: false
          description: Filter to a single FLM by name.
          schema:
            type: string
        - name: reps
          in: query
          required: false
          description: Comma-separated list of rep names to include.
          schema:
            type: string
        - name: regions
          in: query
          required: false
          description: Comma-separated list of regions to include.
          schema:
            type: string
      responses:
        "200":
          description: Goals table rows
          content:
            application/json:
              schema:
                type: object
                required: [month, rows]
                properties:
                  month:
                    type: string
                  rows:
                    type: array
                    items:
                      $ref: "#/components/schemas/GoalTableRow"
  /sales/goals/table/override:
    put:
      operationId: upsertGoalRowOverride
      tags: [sales]
      summary: Upsert a single Goals-table row override
      description: |
        Inserts or updates the per-(month, rep, product) override. Only the
        fields present in the body are changed; omitted fields keep their stored
        value or the code default. Restricted to admin / SLM / exec.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalRowOverrideInput"
      responses:
        "200":
          description: Override saved
          content:
            application/json:
              schema:
                type: object
                required: [override]
                properties:
                  override:
                    $ref: "#/components/schemas/GoalRowOverride"
        "400":
          $ref: "#/components/responses/GoalsValidationError"
  /sales/goals/table/bulk-source:
    post:
      operationId: bulkSetGoalRowSource
      tags: [sales]
      summary: Bulk-set the Source for every row in the filter set
      description: |
        Sets the goal Source for all (rep, product) rows in the current filter
        set (`month` + the optional narrowing filters supplied as query params),
        preserving any other stored per-row override values. Restricted to
        admin / SLM / exec.
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
        - name: slm
          in: query
          required: false
          schema:
            type: string
        - name: flm
          in: query
          required: false
          schema:
            type: string
        - name: reps
          in: query
          required: false
          schema:
            type: string
        - name: regions
          in: query
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source:
                  $ref: "#/components/schemas/GoalSourceId"
      responses:
        "200":
          description: Bulk source applied
          content:
            application/json:
              schema:
                type: object
                required: [month, updated, source]
                properties:
                  month:
                    type: string
                  updated:
                    type: integer
                  source:
                    $ref: "#/components/schemas/GoalSourceId"
        "400":
          $ref: "#/components/responses/GoalsValidationError"
...
  /sales/roster:
    get:
      operationId: getRoster
      tags: [sales]
      summary: List the editable roster for a month
      description: |
        Every person in the base hierarchy (active AND inactive) with their
        base hierarchy values, the per-month overrides, and the resolved
        effective values. Executive-view roles only.
      parameters:
        - name: month
          in: query
          required: false
          description: Month in `YYYY-MM` format. Defaults to the current month.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        "200":
          description: Roster rows for the month
          content:
            application/json:
              schema:
                type: object
                required: [month, rows]
                properties:
                  month:
                    type: string
                  rows:
                    type: array
                    items:
                      $ref: "#/components/schemas/RosterPersonRow"
        "400":
          description: Invalid `month` parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  schemas:
...
    GoalCsvJoinField:
      type: object
      required: [csv, hierarchy]
      description: A join pair linking a Goal CSV column to a hierarchy field.
      properties:
        csv:
          type: string
        hierarchy:
          $ref: "#/components/schemas/HierarchyJoinField"
    FinancePpsJoinField:
      type: object
      required: [financePps, hierarchy]
      description: A join pair linking a finance.pps column to a hierarchy field.
      properties:
        financePps:
          type: string
        hierarchy:
          $ref: "#/components/schemas/HierarchyJoinField"
...
    SoftwarePctRules:
      type: object
      required: [subSource, columnMapping, percentages]
      description: |
        Software % Rules config: the sub-source choice, its column mapping, and
        the per-software-product percentages (whole integers summing to 100).
        Software products are Showcase, Zillow Pro, Follow Up Boss, ZMX.
      properties:
        subSource:
          type: string
          enum: [financePps, goalCsv]
        columnMapping:
          type: object
          required: [financePps, goalCsv]
          properties:
            financePps:
              $ref: "#/components/schemas/SoftwareColumnMap"
            goalCsv:
              $ref: "#/components/schemas/SoftwareColumnMap"
        percentages:
          type: object
          required: [Showcase, Zillow Pro, Follow Up Boss, ZMX]
          properties:
            Showcase:
              type: number
            Zillow Pro:
              type: number
            Follow Up Boss:
              type: number
            ZMX:
              type: number
...
    ResolvedGoals:
      type: object
      description: |
        Resolved goals for a month: rep name → product name → metrics. Products
        and metrics a source does not populate are absent (read as zero
        downstream).
      additionalProperties:
        type: object
        additionalProperties:
          $ref: "#/components/schemas/GoalMetrics"
    GoalSourceId:
      type: string
      description: Which goal source drives a row's base goals.
      enum: [financePps, goalCsv, softwareGnr, softwareAcq]
...
  responses:
    GoalsValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: string
...